Opt: abstract CombatUI class
